requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.1
validators==0.20.0
fpdf==1.7.2
markdown==3.5.2
google-generativeai==0.3.2
python-dotenv==1.0.1
simplejson==3.19.1
dataclasses==0.8
